use backend-specific HandleSQLError implementations
