+++ pkg/storage/mysql/mysql_test.go
@@ -2,10 +2,12 @@ package mysql
 
 import (
 	"context"
+	"database/sql"
 	"testing"
 	"time"
 
 	sq "github.com/Masterminds/squirrel"
+	mysqldriver "github.com/go-sql-driver/mysql"
 	"github.com/oklog/ulid/v2"
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 	"github.com/stretchr/testify/require"
@@ -72,7 +74,7 @@ func TestReadEnsureNoOrder(t *testing.T) {
 			thirdTuple := tuple.NewTupleKey("doc:object_id_3", "relation", "user:user_3")
 
 			err = sqlcommon.Write(ctx,
-				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 				store,
 				[]*openfgav1.TupleKeyWithoutCondition{},
 				[]*openfgav1.TupleKey{firstTuple},
@@ -81,7 +83,7 @@ func TestReadEnsureNoOrder(t *testing.T) {
 
 			// Tweak time so that ULID is smaller.
 			err = sqlcommon.Write(ctx,
-				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 				store,
 				[]*openfgav1.TupleKeyWithoutCondition{},
 				[]*openfgav1.TupleKey{secondTuple},
@@ -90,7 +92,7 @@ func TestReadEnsureNoOrder(t *testing.T) {
 
 			// Tweak time so that ULID is smaller.
 			err = sqlcommon.Write(ctx,
-				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 				store,
 				[]*openfgav1.TupleKeyWithoutCondition{},
 				[]*openfgav1.TupleKey{thirdTuple},
@@ -174,7 +176,7 @@ func TestCtxCancel(t *testing.T) {
 			thirdTuple := tuple.NewTupleKey("doc:object_id_3", "relation", "user:user_3")
 
 			err = sqlcommon.Write(ctx,
-				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 				store,
 				[]*openfgav1.TupleKeyWithoutCondition{},
 				[]*openfgav1.TupleKey{firstTuple},
@@ -183,7 +185,7 @@ func TestCtxCancel(t *testing.T) {
 
 			// Tweak time so that ULID is smaller.
 			err = sqlcommon.Write(ctx,
-				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 				store,
 				[]*openfgav1.TupleKeyWithoutCondition{},
 				[]*openfgav1.TupleKey{secondTuple},
@@ -192,7 +194,7 @@ func TestCtxCancel(t *testing.T) {
 
 			// Tweak time so that ULID is smaller.
 			err = sqlcommon.Write(ctx,
-				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+				sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 				store,
 				[]*openfgav1.TupleKeyWithoutCondition{},
 				[]*openfgav1.TupleKey{thirdTuple},
@@ -235,7 +237,7 @@ func TestReadPageEnsureOrder(t *testing.T) {
 	secondTuple := tuple.NewTupleKey("doc:object_id_2", "relation", "user:user_2")
 
 	err = sqlcommon.Write(ctx,
-		sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+		sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 		store,
 		[]*openfgav1.TupleKeyWithoutCondition{},
 		[]*openfgav1.TupleKey{firstTuple},
@@ -244,7 +246,7 @@ func TestReadPageEnsureOrder(t *testing.T) {
 
 	// Tweak time so that ULID is smaller.
 	err = sqlcommon.Write(ctx,
-		sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()")),
+		sqlcommon.NewDBInfo(ds.db, ds.stbl, sq.Expr("NOW()"), HandleSQLError),
 		store,
 		[]*openfgav1.TupleKeyWithoutCondition{},
 		[]*openfgav1.TupleKey{secondTuple},
@@ -427,4 +429,34 @@ func TestMarshalledAssertions(t *testing.T) {
 		},
 	}
 	require.Equal(t, expectedAssertions, assertions)
+}
+
+func TestHandleSQLError(t *testing.T) {
+	t.Run("duplicate_entry_value_error_with_tuple_key_wraps_ErrInvalidWriteInput", func(t *testing.T) {
+		duplicateKeyError := &mysqldriver.MySQLError{
+			Number:  1062,
+			Message: "Duplicate entry '' for key ''",
+		}
+		err := HandleSQLError(duplicateKeyError, &openfgav1.TupleKey{
+			Object:   "object",
+			Relation: "relation",
+			User:     "user",
+		})
+		require.ErrorIs(t, err, storage.ErrInvalidWriteInput)
+	})
+
+	t.Run("duplicate_entry_value_error_without_tuple_key_returns_collision", func(t *testing.T) {
+		duplicateKeyError := &mysqldriver.MySQLError{
+			Number:  1062,
+			Message: "Duplicate entry '' for key ''",
+		}
+		err := HandleSQLError(duplicateKeyError)
+
+		require.ErrorIs(t, err, storage.ErrCollision)
+	})
+
+	t.Run("sql.ErrNoRows_is_converted_to_storage.ErrNotFound_error", func(t *testing.T) {
+		err := HandleSQLError(sql.ErrNoRows)
+		require.ErrorIs(t, err, storage.ErrNotFound)
+	})
 }
